Switch configuration file format to YAML (#310)
Switch configuration file format to YAML

* Fix example configuration setting

* Change HTTP frontend enabling flag handling

* Change backend drivers init error message format

+++ src/facette/connector/connector.go
@@ -8,7 +8,7 @@ import (
 	"facette/plot"
 )
 
-const connectorDefaultTimeout int64 = 10
+const connectorDefaultTimeout int = 10
 
 var (
 	version string
