+++ src/facette/connector/influxdb.go
@@ -35,7 +35,7 @@ type influxDBMapEntry struct {
 type influxdbConnector struct {
 	name          string
 	url           string
-	timeout       int64
+	timeout       int
 	allowInsecure bool
 	username      string
 	password      string
